requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
Pillow
python-decouple
djangorestframework-simplejwt